Recover 802736D4 keymap fns: dataRemoveKeymap 89.2->100% (musyx synthdata.c form - sid != key operand order, cached long n count avoids post-sth reload, tab[j-1]=tab[j] copy loop), dataInsertKeymap 83.6->99.2% (DATA_KEYMAP_TAB macro over LOCAL hoisted base u8* - base in saved r29 before sndBegin + flat pointer-arith form folds +0x4600 into loop walker IV init where struct-overlay member form kept it as displacement)

+++ include/main/unknown/autos/placeholder_802736D4.h
@@ -8,6 +8,6 @@ void doNothing_802737E8(void);
 void synthCancelJob(int voice);
 void synthRefreshJobVolumes(void);
 int dataInsertKeymap(u16 keymapId, void *data);
-int dataRemoveKeymap(u32 keymapId);
+int dataRemoveKeymap(u16 keymapId);
 
 #endif /* MAIN_UNKNOWN_AUTOS_PLACEHOLDER_802736D4_H_ */
